Cleaning up warnings made by c++

+++ KKBase/SimpleCompressor.cpp
@@ -190,8 +190,8 @@ uchar*  SimpleCompressor::CreateCompressedBuffer (kkuint32&  compressedBuffserSi
   uchar*  compressedBuff = new uchar[compressedBytesNeeded];
   kkuint32  compressedBuffUsed = 0;
 
-  AddByteToCmpressedBuffer (compressedBuff, compressedBuffUsed, compressedBytesNeeded / 256);
-  AddByteToCmpressedBuffer (compressedBuff, compressedBuffUsed, compressedBytesNeeded % 256);
+  AddByteToCmpressedBuffer (compressedBuff, compressedBuffUsed, (uchar)(compressedBytesNeeded / 256));
+  AddByteToCmpressedBuffer (compressedBuff, compressedBuffUsed, (uchar)(compressedBytesNeeded % 256));
 
   kkuint32  idx = 0;
 
@@ -199,7 +199,7 @@ uchar*  SimpleCompressor::CreateCompressedBuffer (kkuint32&  compressedBuffserSi
   {
     while  ((idx < buffSpaceUsed)  &&  (buffLens[idx] > 1))
     {
-      AddByteToCmpressedBuffer (compressedBuff, compressedBuffUsed, 128 + buffLens[idx]);
+      AddByteToCmpressedBuffer (compressedBuff, compressedBuffUsed, (uchar)(128 + buffLens[idx]));
       AddByteToCmpressedBuffer (compressedBuff, compressedBuffUsed, buffBytes[idx]);
       idx++;
     }
@@ -218,9 +218,8 @@ uchar*  SimpleCompressor::CreateCompressedBuffer (kkuint32&  compressedBuffserSi
       }
 
       endIdx--;
-
       
-      AddByteToCmpressedBuffer (compressedBuff, compressedBuffUsed, 0 + rawBytesInARow);
+      AddByteToCmpressedBuffer (compressedBuff, compressedBuffUsed, (uchar)(0 + rawBytesInARow));
       for  (kkuint32 x = 0;   x < rawBytesInARow;  x++)
       {
         AddByteToCmpressedBuffer (compressedBuff, compressedBuffUsed, buffBytes[idx]);
